Clean up Resource Processing
This is an intermediate commit and is not complete.

Changes:
- Adjust the Waste process so that there is a minumum amount before scoring
- Adjust the scoring for the Toggle off processes
- ProcessSpec calculates the mimimum input resources based on 25% of the process time
- ResourceProcess has a new State that identified if the process ig running, idle or out of inputs
- Update the UI so a tri-colour process column is used
- Move all Resource Porcess related classes to a new package
- Collapose the unused ProcessSpec & associated config
- Create new unit test for ResourceProcessConfig

Part of #1513
