+++ mars-sim-core/src/main/java/com/mars_sim/core/SimulationConfig.java
@@ -34,12 +34,12 @@ import com.mars_sim.core.person.health.MedicalConfig;
 import com.mars_sim.core.resource.AmountResourceConfig;
 import com.mars_sim.core.resource.PartConfig;
 import com.mars_sim.core.resource.PartPackageConfig;
+import com.mars_sim.core.resourceprocess.ResourceProcessConfig;
 import com.mars_sim.core.robot.RobotConfig;
 import com.mars_sim.core.science.ScienceConfig;
 import com.mars_sim.core.structure.SettlementConfig;
 import com.mars_sim.core.structure.building.BuildingConfig;
 import com.mars_sim.core.structure.building.BuildingPackageConfig;
-import com.mars_sim.core.structure.building.ResourceProcessConfig;
 import com.mars_sim.core.structure.building.function.cooking.MealConfig;
 import com.mars_sim.core.structure.building.function.farming.CropConfig;
 import com.mars_sim.core.structure.construction.ConstructionConfig;
@@ -70,7 +70,7 @@ public class SimulationConfig {
 	private static final String CROP_FILE = "crops";
 	private static final String LANDMARK_FILE = "landmarks";
 	private static final String MINERAL_MAP_FILE = "minerals";
-	private static final String RESPROCESS_FILE = "resource_process";
+	public static final String RESPROCESS_FILE = "resource_process";
 	private static final String BUILDING_FILE = "buildings";
 	private static final String PART_FILE = "parts";
 	private static final String PART_PACKAGE_FILE = "part_packages";
