+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -50,6 +50,7 @@ import com.mars_sim.core.person.health.task.RequestMedicalTreatmentMeta;
 import com.mars_sim.core.person.health.task.RestingMedicalRecoveryMeta;
 import com.mars_sim.core.person.health.task.SelfTreatHealthProblemMeta;
 import com.mars_sim.core.person.health.task.TreatMedicalPatientMeta;
+import com.mars_sim.core.resourceprocess.task.ToggleResourceProcessMeta;
 import com.mars_sim.core.robot.ai.task.ChargeMeta;
 import com.mars_sim.core.robot.ai.task.SavePowerMeta;
 import com.mars_sim.core.science.task.AssistScientificStudyResearcherMeta;
@@ -70,7 +71,6 @@ import com.mars_sim.core.structure.building.function.task.TendAlgaePondMeta;
 import com.mars_sim.core.structure.building.function.task.TendFishTankMeta;
 import com.mars_sim.core.structure.building.function.task.TendGreenhouseMeta;
 import com.mars_sim.core.structure.building.function.task.ToggleFuelPowerSourceMeta;
-import com.mars_sim.core.structure.building.function.task.ToggleResourceProcessMeta;
 import com.mars_sim.core.structure.building.task.MaintainBuildingMeta;
 import com.mars_sim.core.structure.task.BudgetResourcesMeta;
 import com.mars_sim.core.structure.task.DigLocalIceMeta;
