+++ mars-sim-core/src/main/java/com/mars_sim/core/structure/building/BuildingConfig.java
@@ -23,6 +23,8 @@ import org.jdom2.Element;
 
 import com.mars_sim.core.configuration.ConfigHelper;
 import com.mars_sim.core.map.location.LocalPosition;
+import com.mars_sim.core.resourceprocess.ResourceProcessConfig;
+import com.mars_sim.core.resourceprocess.ResourceProcessEngine;
 import com.mars_sim.core.science.ScienceType;
 import com.mars_sim.core.structure.building.function.FunctionType;
