+++ mars-sim-core/src/main/java/com/mars_sim/core/structure/building/function/ResourceProcessor.java
@@ -9,11 +9,12 @@ package com.mars_sim.core.structure.building.function;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.mars_sim.core.resourceprocess.ResourceProcess;
+import com.mars_sim.core.resourceprocess.ResourceProcessEngine;
 import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.structure.building.Building;
 import com.mars_sim.core.structure.building.BuildingException;
 import com.mars_sim.core.structure.building.FunctionSpec;
-import com.mars_sim.core.structure.building.ResourceProcessEngine;
 import com.mars_sim.core.structure.building.utility.power.PowerMode;
 import com.mars_sim.core.time.ClockPulse;
 import com.mars_sim.core.time.MarsTime;
